+++ src/io/__init__.py
@@ -1,0 +1,5 @@
+"""Módulo de I/O / IO Module."""
+
+from .dascore import alakoro_to_patch, patch_to_alakoro
+
+__all__ = ["alakoro_to_patch", "patch_to_alakoro"]